Rename module to prevent import conflicts with builtin test
